server/requirements.txt
fastapi
uvicorn
deta
python_dotenv